LICENSE
License Protecting the Software and Third-Party Components
This software and the third-party components used are protected by their respective licenses. The use, modification, distribution, or commercial exploitation of CorAlaska Game Studios' software is prohibited without prior authorization. Third-party resources are subject to their own licenses, such as the MIT License, and must be used with proper credit to the author.